msm8998-common: power: Release launch boost perflock when launch is completed
Currently on perf HAL platforms, the launch boost perflock is held for
the fixed duration specified in perfboostsconfig.xml.

Using a fixed duration for this perflock has two shortcomings:
* perflock can be held for too long, causing the CPU freq and other resources
  to be boosted for too long even if the application has finished launching,
  resulting in excessive battery drainage
* perflock can be held for not long enough if it takes more than the timeout
  time to finish launch

The framework sends out a powerhint for when launch starts and when launch
finishes. The launch finish hint can therefore be used to signal when to
release the perflock.

Reference: Wahoo power-libperfmgr

Change-Id: I405cc453c5f58d9fb2583b9c6017f3964a0ce024